penis

+++ col.cpp
@@ -1,8 +1,9 @@
 #include "col.hpp"
 
-Col::Col(const u8 r, const u8 g, const u8 b) 
+Col::Col(const uint r, const uint g, const uint b) 
     : r(r), g(g), b(b) {}
 
 u32 Col::val(const uint roff, const uint goff, const uint boff) const {
+    // TODO: Check for overflow.
     return (r << roff) | (g << goff) | (b << boff);
 }
