+++ col.hpp
@@ -3,7 +3,7 @@
 class Fb;
 class Col {
 public:
-    Col(const u8 r, const u8 g, const u8 b);
+    Col(const uint r, const uint g, const uint b);
     u32 val(const uint roff, const uint goff, const uint boff) const;
 private:
     uint r,g,b;
